Add logic to determine the number of bytes read
Previously, the number of bytes read each time was 128 (constant). This
number was determined by trial and error, for performance
considerations. This required DIM_1 (the x dimension of the images) to
be divisible by (128/(image depth)). Such an approach was based on the
assumption that the dimensions of an image are 2**n, n>6 (which is true
for most EDF images). An error message was printed for other images.

Currently, the minimal number of bytes read is 128 (or less, if the
dimensions of an image are too small for that) and if this does not
satisfy the requirement, this number will go up, until it will satisfy
it.

Signed-off-by: Leonid Bloch <leonid.bloch@esrf.fr>